LLM-494: fix CodeQL uint64→int64 conversion guard recognition
CodeQL flagged int64(filter.LedgerID) as an unchecked narrowing conversion.
The upper-bound guard was there but compared uint64(filter.LedgerID), a
different expression node than the converted operand, so the barrier-guard
analysis didn't link them. Compare filter.LedgerID directly (it's already a
uint64-kinded sim.LedgerID) so the guard and the conversion share the operand.
No behavior change.

+++ engine/sim/repo/pg/action_log.go
@@ -459,7 +459,7 @@ func (r *ActionLogRepo) LoadSettlements(ctx context.Context, filter sim.Settleme
 		// a value above bigint's range cannot exist in the column, so no settlement
 		// can match — return empty rather than bind a wrapped negative value. Real
 		// ids come from a sequence and are nowhere near that bound.
-		if uint64(filter.LedgerID) > math.MaxInt64 {
+		if filter.LedgerID > math.MaxInt64 {
 			return []sim.SettlementRow{}, nil
 		}
 		add(" AND ledger_id = $%d", int64(filter.LedgerID))
